Add optimised keccak reference implementation.
Adding this to the library isn't the cleanest thing ever, but there's no
shared library as far as I can tell, and the only other option I know of
is OpenSSL, which is worse. Still, the reference implementation is
public domain and is pretty small, so no harm in including it.