Initial work on praef_system.
Ended up going with a monolithic approach anyway. Dynamics began to feel
that it would only lead to object hell, and would have left way too much
in implicit and unenforced relationships, not to mention the possibility
of weird issues due to ordering dependencies.

This will mean that the application interface will be more more complex,
but also easier, than with dynamics, as it essentially reduces down to a
table of callbacks. This also makes it easier to be backward-compatible.

Will get rid of the dynamic stuff once it's certain it won't be used. It
is a bit scary, really.