Add standard system bridge.
Not at all tested yet, not that it has that much logic. We'll also
definitely want a higher-level layer of "just give me a fully-assembled
system" to make the library easier to use for applications that don't
need to do anything unusual, but this glue layer is necessary either
way.