Implement interactive TUI menu.
Still needs mouse support, but that can wait until later.

The text input doesn't work as well as it could, but I don't think there
will be high expectations since it's closely emulating a VGA terminal
anyway, and in any case I don't have that much time to spend on this.